chore: Rename Command Center to Terminal in code and views
