+++ resources/views/components/navbar.blade.php
@@ -228,9 +228,9 @@
                         </a>
                     </li>
                     <li>
-                        <a title="Command Center"
-                            class="{{ request()->is('command-center*') ? 'menu-item-active menu-item' : 'menu-item' }}"
-                            href="{{ route('command-center') }}">
+                        <a title="Terminal"
+                            class="{{ request()->is('terminal*') ? 'menu-item-active menu-item' : 'menu-item' }}"
+                            href="{{ route('terminal') }}">
                             <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"
                                 stroke-width="1.5" stroke="currentColor" fill="none" stroke-linecap="round"
                                 stroke-linejoin="round">
@@ -238,7 +238,7 @@
                                 <path d="M5 7l5 5l-5 5" />
                                 <path d="M12 19l7 0" />
                             </svg>
-                            Command Center
+                            Terminal
                         </a>
                     </li>
                     <li>
