+++ routes/web.php
@@ -6,7 +6,6 @@ use App\Http\Controllers\OauthController;
 use App\Http\Controllers\UploadController;
 use App\Livewire\Admin\Index as AdminIndex;
 use App\Livewire\Boarding\Index as BoardingIndex;
-use App\Livewire\CommandCenter\Index as CommandCenterIndex;
 use App\Livewire\Dashboard;
 use App\Livewire\Dev\Compose as Compose;
 use App\Livewire\ForcePasswordReset;
@@ -64,6 +63,7 @@ use App\Livewire\Tags\Show as TagsShow;
 use App\Livewire\Team\AdminView as TeamAdminView;
 use App\Livewire\Team\Index as TeamIndex;
 use App\Livewire\Team\Member\Index as TeamMemberIndex;
+use App\Livewire\Terminal\Index as TerminalIndex;
 use App\Livewire\Waitlist\Index as WaitlistIndex;
 use App\Models\GitlabApp;
 use App\Models\PrivateKey;
@@ -153,11 +153,12 @@ Route::middleware(['auth', 'verified'])->group(function () {
         Route::get('/admin', TeamAdminView::class)->name('team.admin-view');
     });
 
-    Route::get('/command-center', CommandCenterIndex::class)->name('command-center');
+    Route::get('/terminal', TerminalIndex::class)->name('terminal');
     Route::post('/terminal/auth', function () {
         if (auth()->check()) {
             return response()->json(['authenticated' => true], 200);
         }
+
         return response()->json(['authenticated' => false], 401);
     })->name('terminal.auth');
